Refactor AbstractIssueSelector and DefaultIssueSelector for clarity and reuse
